ai-services/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.3.2
moviepy==1.0.3
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.26.2
pydantic==2.5.2
python-multipart==0.0.6
gunicorn==21.2.0
opencv-python==4.9.0.80 